factor UserBehavior out, use it to create the second test plan, add a convenience to run.sh for running either eob-only or all endpoints tests
