+++ locust/Dockerfiles/Dockerfile.tkns
@@ -1,5 +1,5 @@
 FROM golang:1.10
-RUN go get -u github.com/rnagle/go-bluebutton/cmd/tkns
+RUN go get -u github.com/whytheplatypus/go-bluebutton/cmd/tkns
 
 FROM selenium/node-firefox
 COPY --from=0 /go/bin/tkns /usr/local/bin/tkns
