+++ locust/locustfiles/bb/__init__.py
@@ -1,4 +1,3 @@
-#!/usr/bin/python
 """
 BlueButton (2.0!) Locust Test
 
@@ -14,7 +13,7 @@ Environment variables:
     - BB_CLIENT_SECRET:      The client secret of the OAuth application
 """
 
-from locust import HttpLocust, TaskSet, task, web, events
+from locust import TaskSet, web, events
 from requests_oauthlib import OAuth2Session
 import os
 import json
@@ -31,6 +30,7 @@ class UserBehavior(TaskSet):
     count = 0
     max_count = 50
     tokens = []
+    patients = {}
 
     def __init__(self, parent):
         super(UserBehavior, self).__init__(parent)
@@ -65,23 +65,14 @@ class UserBehavior(TaskSet):
     def get_random_userinfo(self):
         access_token = self.get_token()
         self.set_auth_header(access_token)
-        response = self.client.get("%s%s" % (base_url, '/v1/connect/userinfo'))
-        data = json.loads(response.content)
-        return data.get('patient')
+        try:
+            return self.patients[access_token]
+        except KeyError:
+            response = self.client.get("%s%s" % (base_url, '/v1/connect/userinfo'))
+            self.patients[access_token] = json.loads(response.content).get('patient')
+            return self.patients[access_token]
 
     def make_req(self, resource):
         self.check_count()
         self.client.get("%s%s" % (base_url, resource))
         self.record_req()
-
-    @task()
-    def get_eob(self):
-        patient = self.get_random_userinfo()
-        resource = '/v1/fhir/ExplanationOfBenefit/?patient=%s' % patient
-        self.make_req(resource)
-
-
-class WebsiteUser(HttpLocust):
-    task_set = UserBehavior
-    min_wait = int(os.environ.get("BB_LOAD_TEST_MIN_WAIT", 1000))
-    max_wait = int(os.environ.get("BB_LOAD_TEST_MAX_WAIT", 5000))
