+++ locust/run.sh
@@ -1,7 +1,5 @@
 #!/bin/bash
 
-set -e
-
 ###
 # Usage:
 #
@@ -21,8 +19,20 @@ set -e
 # BB_LOAD_TEST_MAX_WAIT (how many ms to wait between requests, upper bound, default: 5000)
 ###
 
+set -e
+
+BB_LOAD_TEST_TYPE="${1:-BB_LOAD_TEST_TYPE}"
+
+if [ "$BB_LOAD_TEST_TYPE" != "eob" ] && [ "$BB_LOAD_TEST_TYPE" != "all" ]
+then
+  echo "Must specify a load test type (eob|all)" >&2
+  exit 1
+fi
+
 docker build -f ./Dockerfiles/Dockerfile.tkns -t bb_tkns .
 docker build -f ./Dockerfiles/Dockerfile.locust -t bb_locust .
+
+set +e
 
 echo "Get access tokens..."
 docker run --rm -it bb_tkns \
@@ -30,6 +40,14 @@ docker run --rm -it bb_tkns \
   -secret $BB_CLIENT_SECRET \
   -url https://${BB_SUB_DOMAIN} \
   -n ${BB_NUM_BENES:-4} > tkns.txt
+
+if [ -z "$(cat tkns.txt)" ]
+then
+  echo "There was a problem fetching access tokens" >&2
+  exit 1
+fi
+
+set -e
 
 echo "Run locust tests..."
 docker run \
@@ -44,6 +62,7 @@ docker run \
   --host https://${BB_SUB_DOMAIN} \
   --no-web \
   --only-summary \
+  -f locustfiles/${BB_LOAD_TEST_TYPE}.py \
   -c ${BB_LOAD_TEST_CONCURRENCY:-1} \
   -r ${BB_LOAD_TEST_HATCH_RATE:-1} \
   -t ${BB_LOAD_TEST_DURATION:-20}
